Put response on single line for predictable output
The test sometimes expects an additional newline at the end of the
output, and other times it doesn't. It seems this is not a direct result
of changes in code that are related to the test.

By using unordered output and putting all entries on a single line, the
test becomes more predictable in whether it will or will not succeed.