Remove plugins
The current understanding of properly ensuring plugin support without
having a hard dependency on a set version of the program, is too limited
for the purposes of having them in the first place.

While the application in itself is still extendable, the plugin system
is removed. It may come back in the future when a better understanding
of shipping Go plugins is available.